+++ src/services/functions.py
@@ -1,21 +1,11 @@
 import os
 import pygame
 import pygame.locals
+from entities.sudoku_grid import SudokuGrid
 
 pygame.font.init()
 dirname = os.path.dirname(__file__)
 
-grid1_values = [
-            [False, False, False, False, False, 2, 1, False, False],
-            [False, False, False, False, 8, False, 6, 7, False],
-            [False, False, 9, False, 1, False, 4, 8, False],
-            [False, 2, False, False, False, False, False, False, False],
-            [6, 9, 8, False, 7, False, False, False, 4],
-            [False, 4, 7, 1, 5, 6, 9, False, False],
-            [False, 7, False, False, 6, 1, 2, False, False],
-            [False, False, False, False, False, False, 8, 4, 6],
-            [2, False, 5, False, 4, False, False, 9, 1]
-        ]
 
 number_font = pygame.font.SysFont("cambria", 25)
 
@@ -27,22 +17,60 @@ class Functions:
         self.textrect = None
         self.textobj = None
         self.error = False
+        self.grids = False
+        self.grid = False
+        self.grid_values = False
+
+    def build_sudoku(self, difficulty):
+        """Valitsee pelattavan Sudokun.
+
+        Args:
+            difficulty: Valittu vaikeustaso
+
+        Returns:
+            self.grid, self.grid_values: Tuple, joka sisältää valitun Sudoku-ruudukon ja sen False-arvot
+        """
+        self.grids = SudokuGrid()
+        if difficulty == 0:
+            self.grid = self.grids.number_grid_easy_1
+            self.grid_values = self.grids.number_grid_easy_1_values
+        elif difficulty == 1:
+            self.grid = self.grids.number_grid_normal_1
+            self.grid_values = self.grids.number_grid_normal_1_values
+        elif difficulty == 2:
+            self.grid = self.grids.number_grid_hard_1
+            self.grid_values = self.grids.number_grid_hard_1_values
+
+        return self.grid, self.grid_values
 
     def draw_text(self, text, font, color, surface, x, y):
+        """Piirtää parametrissä annetun Pygame-tekstin.
+        Args:
+            text: Haluttu teksti.
+            font: Fontti.
+            color: Tekstin väri.
+            surface: Ruutu, johon teksti piirretään
+            x, y: Tekstin koordinaatit
+        """
         self.textobj = font.render(text, 1, color)
         self.textrect = self.textobj.get_rect()
         self.textrect.topleft = (x, y)
         surface.blit(self.textobj, self.textrect)
 
-    def draw_numbers(self, grid):
+    def draw_numbers(self, grid, grid_values):
+        """Piirtää ruudukkoon numerot.
 
+        Args:
+            grid: Valittu ruudukko.
+            grid_values: Valitun ruudukon False-arvot
+        """
         row = 0
         offset = 57
         while row < 9:
             col = 0
             while col < 9:
                 output = grid[row][col]
-                if not grid1_values[row][col]:
+                if not grid_values[row][col]:
                     number_text = number_font.render(
                         str(output), True, (255, 0, 0))
                 else:
@@ -55,6 +83,8 @@ class Functions:
             row += 1
 
     def draw_background(self):
+        """Piirtää tyhjän Sudoku-ruudukon.
+        """
         screen.fill((255, 255, 255))
         pygame.draw.rect(screen, (0, 0, 0),
                          pygame.Rect(35, 40, 540, 540), 5)
@@ -68,6 +98,19 @@ class Functions:
             i += 1
 
     def check(self, k, y, x, grid):
+        """Tarkistaa, onko ruudukossa virhe (jolloin täyttöä ei voi jatkaa ennen virheen korjaamista),
+        onko syötetty numero 0 (jolloin ruutu tyhjenee, mikäli siinä oli jokin numero) ja onko pysty- tai vaakarivissä
+        numeroa joka on ristiriidassa syötetyn numeron kanssa.
+
+        Args:
+            k: Syötetty numero.
+            y, x: Ruudun koordinaatit.
+            grid: Täytettävä ruudukko.
+        Returns:
+            None: Palauttaa arvon None muuttujalle self.error.
+            y, x: Tuple joka sisältää sen ruudun, joka on ilmeisen virheellinen.
+            self.check_if_grid_full(grid): Palauttaa tiedon siitä, onko ruudukko täynnä.
+        """
         if self.error and (y, x) != self.error:
             grid[y][x] = ""
             return
@@ -83,6 +126,20 @@ class Functions:
         for i in range(9):
             if i != y and grid[i][x] == k:
                 return y, x
+
+        return self.check_if_grid_full(grid)
+
+    def check_squares(self, y, x, grid):
+        """Tarkistaa, onko ruutuun asetettu numero ristiriidassa samassa 3x3-ruudukossa olevien numeroiden kanssa.
+
+        Args:
+            y, x: Ruudun koordinaatit.
+            grid: Täytettävä ruudukko.
+
+        Returns:
+            y, x: Tuple joka sisältää sen ruudun, joka on ilmeisen virheellinen.
+            self.check_if_grid_full(grid): Palauttaa tiedon siitä, onko ruudukko täynnä.
+        """
 
         if y <= 2 and x <= 2:
             for i in range(3):
@@ -138,6 +195,18 @@ class Functions:
                     if (i, j) != (y, x) and grid[i][j] == grid[y][x] and grid[i][j] != "":
                         return y, x
 
+        return self.check_if_grid_full(grid)
+
+    def check_if_grid_full(self, grid):
+        """Tarkistaa, onko ruudukko täytetty.
+
+        Args:
+            grid: Täytettävä ruudukko.
+
+        Returns:
+            "victory": Asettaa arvon "victory" muuttujalle self.victory
+            None: Asettaa arvon None muuttujalle self.victory
+        """
         a = 0
         for i in grid:
             for j in i:
@@ -146,5 +215,5 @@ class Functions:
 
         if a == 81:
             return "victory"
-
-        return None
+        else:
+            return None