+++ src/ui/gameplay.py
@@ -3,6 +3,8 @@ import pygame, sys, os
 from pygame.locals import *
 from entities.sudoku_grid import SudokuGrid
 from services.functions import Functions
+
+import time
 pygame.font.init()
 dirname = os.path.dirname(__file__)
 
@@ -11,25 +13,71 @@ class Gameplay:
                 self.screen = pygame.display.set_mode((600, 600))
                 self.error = False
                 self.error_font = pygame.font.SysFont("cambria", 10)
+                self.timer_font = pygame.font.SysFont("cambria", 15)
                 self.victory_font = pygame.font.SysFont("cambria", 20)
+                self.return_font = pygame.font.SysFont("cambria", 40)
                 self.grids = SudokuGrid()
                 self.functions = Functions()
-                self.grid1 = self.grids.number_grid1
-                self.grid1_values = self.grids.number_grid1_values
                 self.number_font = pygame.font.SysFont("cambria", 25)
                 self.mainClock = pygame.time.Clock()
                 self.RED = (255, 0, 0)
                 self.GREY = (128, 128, 128)
-                self.grid1true = False
                 self.full = False
+                self.start = False
+                self.elapsed_time = False
+                self.return_rect = pygame.Rect(200, 270, 200, 50)
+                self.timer_rect = pygame.Rect(500, 5, 80, 20)
+                self.mx, self.my = pygame.mouse.get_pos()
+                self.click = False
+                self.session = False
+                self.running = False
 
+        def sure(self):
+                """Varmistaa, haluaako pelaaja keskeyttää sudokun
 
+                Returns:
+                        Palauttaa True, jos painetaan ESC, ja False, jos painetaan Enter
+                """
+                while self.running:
+                        for event in pygame.event.get():
+                                if event.type == QUIT:
+                                        pygame.quit()
+                                        sys.exit()
+                                if event.type == KEYDOWN:
+                                        if event.key == K_ESCAPE:
+                                                return True
+                                        elif event.key == K_RETURN:
+                                                return False
 
-        def game_loop(self):
+                        pygame.draw.rect(self.screen, (255, 255, 255), self.return_rect)
+                        pygame.draw.rect(self.screen, (0, 0, 0), self.return_rect, 1)
+                        self.functions.draw_text('Haluatko varmasti keskeyttää?', self.timer_font, (0, 0, 0),
+                                                 self.screen, 205, 272)
+                        self.functions.draw_text('ESC = Poistu', self.error_font, self.RED, self.screen, 269, 290)
+                        self.functions.draw_text('ENTER = Jatka', self.error_font, self.RED, self.screen, 265, 302)
+                        self.end = time.time()
+                        pygame.draw.rect(self.screen, (255, 255, 255), self.timer_rect)
+                        self.timer(self.start, self.end)
+                        pygame.display.update()
+                        self.mainClock.tick(60)
+
+        def game_loop(self, difficulty):
+                """Sudokun täyttämisen käyttöliittymä. Tarkistaa ensin onko Sudokua ensin rakennettu, sitten käynnistää
+                timerin ja täyttää täyttämättömän ruudukon arvoja kuvaavan dictionaryn.
+
+                Args:
+                        difficulty: Valittu vaikeustaso.
+                """
+                if not self.session:
+                        self.grid = self.functions.build_sudoku(difficulty)[0]
+                        self.grid_values = self.functions.build_sudoku(difficulty)[1]
+                        self.session = True
+
+                if not self.start:
+                        self.start = time.time()
+
                 self.running = True
                 self.click = False
-                self.grid = self.grid1
-                self.grid_values = self.grid1_values
                 x_factor = 60
                 y_factor = 60
                 grid_value_dict = {}
@@ -46,14 +94,24 @@ class Gameplay:
                                         sys.exit()
                                 if event.type == KEYDOWN:
                                         if event.key == K_ESCAPE:
-                                                self.running = False
+                                                if self.sure():
+                                                        self.running = False
+                                                        self.start = False
+                                                        self.end = False
+                                                        self.session = False
+                                                        self.error = False
                                         if event.key >= 48 and event.key <= 57:
                                                 if highlight:
                                                         self.grid[highlight[0]][highlight[1]] = event.key - 48
-                                                        self.error = self.functions.check(self.grid[highlight[0]][highlight[1]], highlight[0],
-                                                                   highlight[1], self.grid)
+                                                        self.error = self.functions.check(
+                                                                self.grid[highlight[0]][highlight[1]], highlight[0],
+                                                                highlight[1], self.grid)
+                                                        if not self.error:
+                                                                self.error = self.functions.check_squares(highlight[0],
+                                                                                                          highlight[1],
+                                                                                                          self.grid)
                                                         if self.error == "victory":
-                                                                self.victory()
+                                                                self.victory(self.elapsed_time)
 
                                 if event.type == MOUSEBUTTONDOWN:
                                         if event.button == 1:
@@ -63,7 +121,7 @@ class Gameplay:
                         self.mx, self.my = pygame.mouse.get_pos()
 
                         self.functions.draw_background()
-                        self.functions.draw_numbers(self.grid)
+                        self.functions.draw_numbers(self.grid, self.grid_values)
 
                         for i in range(9):
                                 for j in range(9):
@@ -84,12 +142,30 @@ class Gameplay:
                         if self.error:
                                 self.functions.draw_text('Väärin!', self.error_font, (255, 0, 0), self.screen, 5, 5)
 
+                        self.end = time.time()
+                        self.timer(self.start, self.end)
                         pygame.display.update()
                         self.mainClock.tick(60)
 
-        def victory(self):
-                self.running = True
+        def timer(self, start, end):
+                """Timerin ulkoasu.
 
+                Args:
+                        start: Lähtöaika.
+                        end: Lopetusaika.
+                """
+                hours, rem = divmod(end - start, 3600)
+                minutes, seconds = divmod(rem, 60)
+                self.elapsed_time = "{:0>2}:{:0>2}:{:05.2f}".format(int(hours), int(minutes), seconds)
+                self.functions.draw_text(self.elapsed_time, self.timer_font, (0, 0, 0), self.screen, 500, 5)
+
+        def victory(self, elapsed):
+                """Voittoruudun käyttöliittymä.
+
+                Args:
+                        elapsed: Ratkaisuun käytetty aika.
+                """
+                self.running = True
                 while self.running:
                         for event in pygame.event.get():
                                 if event.type == QUIT:
@@ -98,9 +174,21 @@ class Gameplay:
                                 if event.type == KEYDOWN:
                                         if event.key == K_ESCAPE:
                                                 self.running = False
+                                                self.session = False
+                                                self.start = False
+                                                self.end = False
+                                                self.error = None
+                                if event.type == MOUSEBUTTONDOWN:
+                                        if event.button == 1:
+                                                self.click = True
 
                         self.functions.draw_background()
                         self.functions.draw_text('Hyvää työtä!', self.victory_font, (0, 0, 0), self.screen, 5, 5)
+                        self.functions.draw_text('Aika: ' + str(elapsed), self.timer_font, (0, 0, 0), self.screen, 150, 5)
+                        pygame.draw.rect(self.screen, (255, 255, 255), self.return_rect)
+                        pygame.draw.rect(self.screen, (0, 0, 0), self.return_rect, 1)
+                        self.functions.draw_text('Palaa painamalla ESC', self.victory_font, (0, 0, 0), self.screen, 210, 282)
+
 
                         pygame.display.update()
                         self.mainClock.tick(60)